Damage taking overhaul
