+++ src/main/java/com/vetpetmon/wyrmsofnyrus/entity/wyrms/EntityWyrmSoldierInfectoid.java
@@ -2,7 +2,6 @@ package com.vetpetmon.wyrmsofnyrus.entity.wyrms;
 
 import com.vetpetmon.wyrmsofnyrus.SoundRegistry;
 import com.vetpetmon.wyrmsofnyrus.config.Evo;
-import com.vetpetmon.wyrmsofnyrus.config.Radiogenetics;
 import com.vetpetmon.wyrmsofnyrus.config.wyrmStats;
 import com.vetpetmon.wyrmsofnyrus.entity.EntityWyrm;
 import com.vetpetmon.wyrmsofnyrus.evo.evoPoints;
@@ -87,14 +86,6 @@ public class EntityWyrmSoldierInfectoid extends EntityWyrm implements IAnimatabl
 
     public boolean attackEntityFrom(DamageSource source, float amount) {
         if (source.isMagicDamage()) amount *= 0.5F;
-        if (source == DamageSource.FALL && Radiogenetics.immuneToFalling)
-            return false;
-        if (source == DamageSource.DROWN)
-            return false;
-        if (source == DamageSource.CACTUS && Radiogenetics.immuneToCacti)
-            return false;
-        if (source == DamageSource.ON_FIRE)
-            return super.attackEntityFrom(source, amount*3);
         return super.attackEntityFrom(source, amount);
     }
