+++ src/main/java/com/vetpetmon/wyrmsofnyrus/entity/wyrms/EntityWyrmWarriorTainted.java
@@ -6,7 +6,6 @@ import com.hbm.potion.HbmPotion;
 import com.vetpetmon.wyrmsofnyrus.SoundRegistry;
 import com.vetpetmon.wyrmsofnyrus.config.Client;
 import com.vetpetmon.wyrmsofnyrus.config.Evo;
-import com.vetpetmon.wyrmsofnyrus.config.Radiogenetics;
 import com.vetpetmon.wyrmsofnyrus.config.wyrmStats;
 import com.vetpetmon.wyrmsofnyrus.entity.EntityWyrmFlying;
 import com.vetpetmon.wyrmsofnyrus.evo.evoPoints;
@@ -124,14 +123,6 @@ public class EntityWyrmWarriorTainted extends EntityWyrmFlying implements IAnima
             this.heal(amount * 1.25F);
             return false;
         }
-        if (source == DamageSource.FALL)
-            return false;
-        if (source == DamageSource.DROWN)
-            return false;
-        if (source == DamageSource.CACTUS && Radiogenetics.immuneToCacti)
-            return false;
-        if (source == DamageSource.ON_FIRE)
-            return super.attackEntityFrom(source, amount*3);
         return super.attackEntityFrom(source, amount);
     }
